+++ astro.config.mjs
@@ -19,6 +19,11 @@ export default defineConfig({
 
   vite: {
     plugins: [tailwindcss()],
+    resolve: {
+      alias: {
+        '@': new URL('./src', import.meta.url).pathname
+      }
+    },
     server: {
       watch: {
         ignored: [
