+++ src/components/layout/header/article-utils.ts
@@ -7,8 +7,8 @@ import {
   generateLanguageUrlsForArticle,
   getLanguageName,
   getLanguageFlag,
-} from "../../i18n/utils";
-import type { Languages } from "../../i18n/ui";
+} from "@/i18n/utils";
+import type { Languages } from "@/i18n/ui";
 import type {
   ArticleLanguageContext,
   LanguageUrls,
