+++ src/components/meta/SEOHead.astro
@@ -1,6 +1,6 @@
 ---
-import { getLangFromUrl, useTranslations } from "../../i18n/utils";
-import { siteConfig, siteUtils } from "../../config/site";
+import { getLangFromUrl, useTranslations } from "@/i18n/utils";
+import { siteConfig, siteUtils } from "@/config/site";
 
 interface Props {
   title?: string;
