+++ src/components/ui/ArticleCard.astro
@@ -1,5 +1,5 @@
 ---
-import { getLangFromUrl, useTranslations } from "../../i18n/utils";
+import { getLangFromUrl, useTranslations } from "@/i18n/utils";
 
 export interface Props {
   title: string;
