+++ src/pages/category/[category].astro
@@ -1,13 +1,13 @@
 ---
-import CategoryPage from "../../components/pages/CategoryPage.astro";
+import CategoryPage from "@/components/pages/CategoryPage.astro";
 import { getCollection } from "astro:content";
 import {
   getUniqueCategories,
   filterPostsByCategory,
   normalizeCategoryForUrl,
   type CategoryTranslations,
-} from "../../lib/article/category-utils";
-import { getCategoryTranslations } from "../../i18n/category-translations";
+} from "@/lib/article/category-utils";
+import { getCategoryTranslations } from "@/i18n/category-translations";
 
 export async function getStaticPaths() {
   // Récupérer tous les articles
