+++ src/pages/fr/index.astro
@@ -1,8 +1,8 @@
 ---
-import MainLayout from "../../layouts/MainLayout.astro";
-import LatestArticlesSection from "../../components/sections/LatestArticlesSection.astro";
-import { getLangFromUrl, useTranslations } from "../../i18n/utils";
-import "../../styles/global/app.css";
+import MainLayout from "@/layouts/MainLayout.astro";
+import LatestArticlesSection from "@/components/sections/LatestArticlesSection.astro";
+import { getLangFromUrl, useTranslations } from "@/i18n/utils";
+import "@/styles/global/app.css";
 
 // Système d'internationalisation
 const lang = getLangFromUrl(Astro.url);
