+++ src/test/layouts/MainLayout.test.ts
@@ -10,16 +10,16 @@ describe("MainLayout Structure and Configuration", () => {
 
   it("should import required dependencies", () => {
     expect(layoutContent).toContain(
-      'import Header from "../components/header/Header.astro"',
+      'import Header from "@/components/layout/header/Header.astro"',
     );
     expect(layoutContent).toContain(
-      'import Footer from "../components/layout/Footer.astro"',
+      'import Footer from "@/components/layout/Footer.astro"',
     );
     expect(layoutContent).toContain(
-      'import { getLangFromUrl, useTranslations } from "../i18n/utils"',
+      'import { getLangFromUrl, useTranslations } from "@/i18n/utils"',
     );
     expect(layoutContent).toContain(
-      'import { siteConfig, siteUtils } from "../config/site"',
+      'import { siteConfig, siteUtils } from "@/config/site"',
     );
   });
